requirements.txt
Django~=4.0.4
requests~=2.27.1
djangorestframework~=3.13.1
ujson~=5.2.0
PyYAML~=6.0
celery~=5.2.6
django-allauth
pytest-django
baker
model_bakery
django_rest_passwordreset